Move cli2 outside of cli package
