+++ cli2/upload.py
@@ -7,7 +7,7 @@ import click
 
 from . import api
 from . import gcloud
-from ..constants import UPLOAD_WORKSPACE
+from .constants import UPLOAD_WORKSPACE
 
 
 def upload_assay(assay_type: str, xlsx_path: str):
